Safari-609.1.11.1/LayoutTests/css3/flexbox/flex-flow-auto-margins-no-available-space.html

<!DOCTYPE html>
<html>
<link href="resources/flexbox.css" rel="stylesheet">
<style>
.container {
    position: relative;
    background-color: pink;
    outline: 1px solid black;
    display: inline-block;
}
.flexbox {
    background-color: grey;
    width: 100px;
    height: 100px;
    margin: 20px;
}
.flexbox > * {
    flex: none;
}
.flexbox > :nth-child(1) {
    background-color: blue;
    margin: auto;
}
</style>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<body onload="checkLayout('.flexbox')">
<div id=log></div>

<div class="container">
  <div class="flexbox row">
    <div data-offset-x=60 data-offset-y=20 style="width: 20px; height: 120px"></div>
  </div>
</div>

<div class="container">
  <div class="flexbox row-reverse">
    <div data-offset-x=60 data-offset-y=20 style="width: 20px; height: 120px"></div>
  </div>
</div>

<div class="container">
  <div class="flexbox column">
    <div data-offset-x=20 data-offset-y=60 style="width: 120px; height: 20px"></div>
  </div>
</div>

<div class="container">
  <div class="flexbox column-reverse">
    <div data-offset-x=20 data-offset-y=60 style="width: 120px; height: 20px"></div>
  </div>
</div>
</body>
</html>
